requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
pyyaml>=6.0
anthropic>=0.44.0
openai>=1.60.0
pytest>=8.3.0
httpx>=0.28.0
